@@ -31,6 +31,7 @@
     "@babel/plugin-transform-runtime": "^7.11.5",
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-stage-2": "^7.0.0",
+    "@vue/compiler-sfc": "^3.0.4",
     "audit-ci": "^2.4.2",
     "autoprefixer": "^9.7.3",
     "babel-core": "^7.0.0-bridge.0",
@@ -73,16 +74,15 @@
     "rollup-plugin-flow-no-whitespace": "^1.0.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-vue": "^5.1.4",
+    "rollup-plugin-vue": "^6.0.0",
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.1",
     "terser": "^4.4.2",
     "tippy.js": "^6.1.0",
-    "vue": "^2.6.12",
+    "vue": "^3.0.0",
     "vue-loader": "^15.7.2",
     "vue-router": "^3.4.3",
     "vue-style-loader": "^4.1.0",
-    "vue-template-compiler": "^2.6.12",
     "webpack": "^4.41.2",
     "webpack-dev-middleware": "^3.7.1",
     "webpack-hot-middleware": "^2.25.0",
