@@ -32,7 +32,6 @@
     "tiptap-utils": "^1.11.0"
   },
   "peerDependencies": {
-    "vue": "^2.5.17",
-    "vue-template-compiler": "^2.5.17"
+    "vue": "^3.0.0"
   }
 }
