@@ -54,7 +54,7 @@ export default {
       return null
     }
 
-    return this.$scopedSlots.default({
+    return this.$slots.default({
       focused: this.editor.view.focused,
       focus: this.editor.focus,
       commands: this.editor.commands,
@@ -65,7 +65,7 @@ export default {
     })
   },
 
-  beforeDestroy() {
+  beforeUnmount() {
     this.editor.unregisterPlugin('menu_bubble')
   },
 
